+++ src/test/utiltest/timestampHoverProvider.test.ts
@@ -51,11 +51,11 @@ describe('TimestampHoverProvider', () => {
             new vscode.Position(3, 32),
             tokenSource.token);
 
-        assert.equal(timeConverterMock.isValidEpoch.calledOnce, true);
-        assert.equal(timeConverterMock.epochToISOUtc.calledOnce, true);
-        assert.equal(timeConverterMock.epochToISOUtc.firstCall.args[0], '123456789000');
-        assert.equal(result.contents.length, 1);
-        assert.equal(
+        assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToISOUtc.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToISOUtc.firstCall.args[0], '123456789000');
+        assert.strictEqual(result.contents.length, 1);
+        assert.strictEqual(
             result.contents[0],
             '*Epoch Unit*: `s`  \n*Timestamp*: `test-time`');
     });
@@ -70,11 +70,11 @@ describe('TimestampHoverProvider', () => {
             new vscode.Position(3, 32),
             tokenSource.token);
 
-        assert.equal(timeConverterMock.isValidEpoch.calledOnce, true);
-        assert.equal(timeConverterMock.epochToIsoLocal.calledOnce, true);
-        assert.equal(timeConverterMock.epochToIsoLocal.firstCall.args[0], '123456789000');
-        assert.equal(result.contents.length, 1);
-        assert.equal(
+        assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToIsoLocal.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToIsoLocal.firstCall.args[0], '123456789000');
+        assert.strictEqual(result.contents.length, 1);
+        assert.strictEqual(
             result.contents[0],
             '*Epoch Unit*: `s`  \n*Timestamp*: `test-time`');
     });
@@ -89,12 +89,12 @@ describe('TimestampHoverProvider', () => {
             new vscode.Position(3, 32),
             tokenSource.token);
 
-        assert.equal(timeConverterMock.isValidEpoch.calledOnce, true);
-        assert.equal(timeConverterMock.epochToCustom.calledOnce, true);
-        assert.equal(timeConverterMock.epochToCustom.firstCall.args[0], '123456789000');
-        assert.equal(timeConverterMock.epochToCustom.firstCall.args[1], 'YYYY');
-        assert.equal(result.contents.length, 1);
-        assert.equal(
+        assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToCustom.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[0], '123456789000');
+        assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[1], 'YYYY');
+        assert.strictEqual(result.contents.length, 1);
+        assert.strictEqual(
             result.contents[0],
             '*Epoch Unit*: `s`  \n*Timestamp*: `test-time`');
     });
@@ -107,8 +107,8 @@ describe('TimestampHoverProvider', () => {
             new vscode.Position(0, 0),
             tokenSource.token);
 
-        assert.equal(result, undefined);
-        assert.equal(timeConverterMock.isValidEpoch.notCalled, true);
+        assert.strictEqual(result, undefined);
+        assert.strictEqual(timeConverterMock.isValidEpoch.notCalled, true);
     });
 
     it('should return undefined if it is disabled.', async () => {
@@ -120,8 +120,8 @@ describe('TimestampHoverProvider', () => {
             new vscode.Position(3, 32),
             tokenSource.token);
 
-        assert.equal(result, undefined);
-        assert.equal(timeConverterMock.isValidEpoch.calledOnce, true);
+        assert.strictEqual(result, undefined);
+        assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
     });
 
     it('should return undefined if config is empty.', async () => {
@@ -133,7 +133,7 @@ describe('TimestampHoverProvider', () => {
             new vscode.Position(3, 32),
             tokenSource.token);
 
-        assert.equal(result, undefined);
-        assert.equal(timeConverterMock.isValidEpoch.calledOnce, true);
+        assert.strictEqual(result, undefined);
+        assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
     });
 });